Allow administrator to specify public host name
This is necessary to allow connection from outside the server to the
service.

The default value has been set to localhost, in order to keep the
current behavior in single machine test setups.

PUBLIC_ADDRESS has been used instaed of HOSTNAME it seems to be
specially handled, with everything after the first dot being stripped
off to generate a non-qualified hostname.