Configurable cache location
By default docker volumes are created in /var/lib/docker/volumes... This
can be an issue when working with large cache values as this is generaly
located on the main system volume.

This changes the default behavior to be bind mounted to the provided
folder.

Notes: This has implication for multi-server deployment, but as this is
       not the target of these deployment scripts (yet), it does not
       matter.