Fix folder access
Bind mounts require the folder to exist before mounting, and unlike
regulare -v source:target, docker/docker-compose will not create the
folder when needed.

If the folder exists, AND is empty, docker/docker-compose will set the
permissions as required by the containers.

Git does not allows us to track empty folder AND does not allow us to
set rights on folders.

The solution, for now at least, is to check for the existance of the
folder, and if not, to create it from `run.sh` script.